⚽️ refactor: update keymap export paths to reflect restructured AI and terminal modules
