+++ src/keymaps/ai/configs.ts
