+++ src/keymaps/index.ts
@@ -1,5 +1,5 @@
-export * from './ai/ai-handler';
-export * from './ai/utils';
-export * from './terminal/left-right';
-export * from './terminal/bottom';
+export * from './ai/controller';
+export * from './ai/configs';
+export * from './terminal/side-panel';
+export * from './terminal/bottom-panel';
 export * from './terminal/shared';
