+++ src/keymaps/terminal/bottom-panel.ts
