+++ src/keymaps/terminal/side-panel.ts
